Move 1d gaussian to testsystems
